+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1
-Minor bug fixes in STDIO server
+- Minor bug fixes in STDIO server
+- More efficient client reads
 
 ## 0.1.0
 First release